refactoring code, adding Card and Button component
