+++ src/App.js
@@ -1,8 +1,8 @@
 import "./App.css";
-import Main from "./content/Main";
-import Footer from "./footer/Footer";
-import Banner from "./header/Banner";
-import Header from "./header/Header";
+import Main from "./components/content/Main";
+import Footer from "./components/footer/Footer";
+import Banner from "./components/header/Banner";
+import Header from "./components/header/Header";
 
 function App() {
   return (
