+++ src/components/images/laptop.jpg
