+++ src/components/images/smartphone.jpg
